Src compile WIP 1

+++ designsystem/src/main/java/com/babylon/wallet/android/designsystem/theme/RadixColors.kt
@@ -83,6 +83,6 @@ val AccountGradientList =
         GradientAccount12,
     )
 
-fun getAccountGradientColorsFor(index: Int): List<Color> {
-    return AccountGradientList[index % AccountGradientList.size]
+fun getAccountGradientColorsFor(index: UByte): List<Color> {
+    return AccountGradientList[index.toInt() % AccountGradientList.size]
 }
